pyproject.toml
[project]
name = "mister-white-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "openai>=1.35.0",
  "python-dotenv>=1.0.1",
  "mistralai>=1.2.4",
  "pydantic>=2.8.2",
  "nikitas-agents @ git+https://github.com/NikitaDmitrieff/nikitas-agents.git",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "websockets>=13.0",
  "httpx>=0.27.0",
  "requests>=2.32.0",
]
